Clean up warnings, make default blockSize bigger; I'd rather have lookups take longer (smaller tree, longer in-node search) than run out of memory by default.
Currently opening a names part with 200 million documents freezes up with ~8G of RAM, at least partially due to 512 blockSize. The default is too low for these parts.

+++ core/src/main/java/org/lemurproject/galago/core/index/disk/DiskBTreeReader.java
@@ -80,8 +80,8 @@ public class DiskBTreeReader extends BTreeReader {
       // now we decode everything from the stream
       this.endValueFileOffset = startFileOffset + blockInfo.length;
       this.keyCount = (int) blockStream.readLong();
-      this.keyCache = new byte[(int) this.keyCount][];
-      this.endValueOffsetCache = new long[(int) this.keyCount];
+      this.keyCache = new byte[this.keyCount][];
+      this.endValueOffsetCache = new long[this.keyCount];
       this.startValueFileOffset = this.startFileOffset + this.blockInfo.headerLength;
       this.keyIndex = 0;
       this.done = false;
